+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsefunction.py
@@ -507,9 +507,9 @@ class EvaluatorResponseFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-21T08:06:06.924Z"
+    created: Optional[str] = "2026-07-21T08:34:12.428Z"
 
-    updated: Optional[str] = "2026-07-21T08:06:06.924Z"
+    updated: Optional[str] = "2026-07-21T08:34:12.428Z"
 
     updated_by_id: OptionalNullable[str] = UNSET