+++ packages/orq-rc/RELEASES.md
@@ -9569,3 +9569,13 @@ Based on:
 - [python v4.12.0-rc.45] packages/orq-rc
 ### Releases
 - [PyPI v4.12.0-rc.45] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.45 - packages/orq-rc
+
+## 2026-07-21 08:36:42
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.3 (2.918.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.12.0-rc.46] packages/orq-rc
+### Releases
+- [PyPI v4.12.0-rc.46] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.46 - packages/orq-rc